Migrate jsh plugin definition JSAPI -> Fifty
