+++ jsh/etc/api.html
@@ -273,7 +273,7 @@ END LICENSE
 				</tr>
 			</thead>
 			<tbody>
-				<tr><td><a href="../../jsh/loader/plugin.api.html">(plugins)</a></td><td>Interface for creating <code>jsh</code> plugins</td><!-- no tests --></tr>
+				<tr><td>(plugins)</td><td>Interface for creating <code>jsh</code> plugins</td><!-- no tests --></tr>
 			</tbody>
 		</table>
 	</div>
