+++ jsh/etc/api.js
@@ -72,7 +72,6 @@
 	components.add("jsh/loader/loader.api.html", { jsh: { api: true } });
 	components.add("rhino/shell/plugin.jsh.api.html", { jsh: { api: true } });
 	components.add("jsh/launcher/", { jsh: { api: true } });
-	components.add("jsh/loader/plugin.api.html", { jsh: { api: true } });
 	components.add("jsh/unit/", { jsh: { api: true, module: true } });
 
 	components.add("jsh/tools/install/", { jsh: { api: true, module: true } });
