+++ jsh/loader/jsh.fifty.ts
@@ -212,45 +212,3 @@ namespace slime.jsh {
 		loader: slime.jsh.loader.Exports
 	}
 }
-
-namespace slime.jsh.plugin {
-	export interface EngineSpecific {
-		//	provided by engine-specific rhino.js and nashorn.js
-		exit: any
-		jsh: any
-	}
-
-	export interface Stdio {
-		getStandardInput(): slime.jrunscript.native.java.io.InputStream
-		getStandardOutput(): slime.jrunscript.native.java.io.PrintStream
-		getStandardError(): slime.jrunscript.native.java.io.PrintStream
-	}
-
-	export interface $slime extends slime.jrunscript.runtime.Exports, EngineSpecific {
-		getSystemProperty(name: string): string
-		getEnvironment(): slime.jrunscript.native.inonit.system.OperatingSystem.Environment
-		getInvocation(): slime.jrunscript.native.inonit.script.jsh.Shell.Invocation
-
-		getPackaged(): slime.jrunscript.native.inonit.script.jsh.Shell.Packaged
-
-		plugins: {
-			mock: slime.jsh.loader.internal.plugins.Export["mock"]
-		}
-
-		loader: slime.jrunscript.runtime.Exports["old"]["loader"]
-			& slime.jrunscript.runtime.Exports["loader"]
-			& {
-				getLoaderScript(path: string): any
-			}
-
-		/**
-		 * Returns a `java.io.File` representing a file location relative to the `jsh` library location.
-		 *
-		 * @param path A relative path.
-		 */
-		getLibraryFile: (path: string) => slime.jrunscript.native.java.io.File
-		getInterface(): any
-		getSystemProperties(): slime.jrunscript.native.java.util.Properties
-		getStdio(): Stdio
-	}
-}
