+++ jsh/unit/plugin.jsh.api.html
@@ -356,7 +356,7 @@ END LICENSE
 																				<div class="name">plugins</div>
 																				<span class="type">object</span>
 																				<span>
-																					(optional) The object to use as the <code><a href="../../jsh/loader/plugin.api.html">plugins</a></code> object
+																					(optional) The object to use as the <code>plugins</code> object
 																					when loading the plugins.
 																				</span>
 																			</li>
